final version
final submitted version

+++ shobhitmaheshwari-mychat-d998d51d11b6/app/src/main/java/com/example/shobhit/mychat/response/MessageDetails.java
@@ -22,7 +22,7 @@ public class MessageDetails {
     public String nickname;
     @SerializedName("id")
     @Expose
-    public int result_id;
+    public long result_id;
 
     }
